requirements.txt
cython==0.29.36
D4RL @ git+https://github.com/Farama-Foundation/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288
dm-control==1.0.10
dm-env==1.6
dm-tree==0.1.8
gym==0.17.2
gymnasium==0.27.0
gymnasium-robotics==1.2.0
hydra-core==1.2.0
imageio-ffmpeg==0.4.9
matplotlib==3.6.2
mujoco==2.3.2
mujoco-py==2.1.2.14
networkx==3.1
numpy==1.22.4
omegaconf==2.3.0
scipy==1.10.1
tensorboardX==2.5.1
torch==2.0.0
tqdm==4.64.1
